// e2e/enums/component-enum/buttons.enums.ts
export enum Buttons {
  SIGNIN_lINK = "link-signin",
  SIGNUP_LINK = "link-signup",
  FORGOT_LINK = "link-forgot-password",
  SIGNIN_BUTTON = "signin",
  SIGNUP_BUTTON = "signup-create-account",
  RESET_BUTTON = "Send-Reset-Link",
  VIEW_BUTTON = "view",
  CREATE_PRODUCT = "create-product",
  ADD_SELECTED_PRODUCT = "add-selected-product",
  CHOOSE_FILES = "choose-files",
  INCREMENT = "increment",
}
export enum MenuItems {
  DASHBOARD = "Dashboard",
  PRODUCTS = "Products",
  INVENTORY = "Inventory",
  CART = "Cart",
  REPORTS = "Reports",
  SUPPORT = "Support",
}